requirements.txt
torch==2.9.1
transformers==4.57.3
spacy==3.8.11
playwright==1.58.0
openai-whisper==20250625
requests==2.32.3
beautifulsoup4==4.13.4

# After installing dependencies run:
#   python -m spacy download en_core_web_sm
#   playwright install chromium
